+++ packages/twenty-server/src/engine/metadata-modules/flat-entity/types/__tests__/metadata-many-to-one-join-column.type-test.ts
@@ -1,12 +1,8 @@
-import { type Expect, type Equal } from 'twenty-shared/testing';
+import { type Equal, type Expect } from 'twenty-shared/testing';
 
 import { type MetadataManyToOneJoinColumn } from 'src/engine/metadata-modules/flat-entity/types/metadata-many-to-one-join-column.type';
 
 type FieldMetadataJoinColumns = MetadataManyToOneJoinColumn<'fieldMetadata'>;
-
-type ViewJoinColumns = MetadataManyToOneJoinColumn<'view'>;
-
-type ObjectMetadataJoinColumns = MetadataManyToOneJoinColumn<'objectMetadata'>;
 
 // eslint-disable-next-line unused-imports/no-unused-vars
 type Assertions = [
@@ -19,18 +15,4 @@ type Assertions = [
       | 'relationTargetObjectMetadataId'
     >
   >,
-
-  // view foreign keys from ALL_METADATA_RELATIONS
-  Expect<
-    Equal<
-      ViewJoinColumns,
-      | 'objectMetadataId'
-      | 'calendarFieldMetadataId'
-      | 'kanbanAggregateOperationFieldMetadataId'
-      | 'mainGroupByFieldMetadataId'
-    >
-  >,
-
-  // objectMetadata has no foreign keys (only workspace/application which are null)
-  Expect<Equal<ObjectMetadataJoinColumns, never>>,
 ];
