+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration/universal-flat-entity/types/universal-flat-field-metadata.type.ts
@@ -1,7 +1,7 @@
 import { type FieldMetadataType } from 'twenty-shared/types';
 
 import { type FieldMetadataEntity } from 'src/engine/metadata-modules/field-metadata/field-metadata.entity';
-import { type UniversalFlatEntityFrom } from 'src/engine/metadata-modules/universal-flat-entity/types/universal-flat-entity-from.type';
+import { type UniversalFlatEntityFrom } from 'src/engine/workspace-manager/workspace-migration/universal-flat-entity/types/universal-flat-entity-from.type';
 
 export type UniversalFlatFieldMetadata<
   T extends FieldMetadataType = FieldMetadataType,
